R89: mypy 344→0 全量清零——153 檔型別修飾＋budget 棘輪鎖定 0
apps/backend/src 688 檔零誤報。大宗模式：
- max(d, key=d.get) overloaded 誤報 → key=lambda k: d[k]
- numpy f32→f64 提升 → .astype(np.float32)；np.savez(**dict) stubs 缺 **kwds → 窄化 ignore
- TypedDict vs dict[str, Any] → helper 簽名統一 HAMDataPackageInternal
- asyncio.get_event_loop() → get_running_loop()；pytest fixture 過寬 mock 收窄

scripts/mypy_budget.txt 同步寫 0，budget gate errors==budget 綠燈鎖定。

+++ apps/backend/src/ai/alignment/adversarial_generation_system.py
@@ -121,7 +121,7 @@ class AdversarialGenerationSystem:
         if not self._evaluation_history:
             return 1.0
         scores = [e["robustness_score"] for e in self._evaluation_history]
-        return round(sum(scores) / len(scores), 3)
+        return float(round(sum(scores) / len(scores), 3))
 
 
 __all__ = ["AdversarialGenerationSystem"]
